# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 퍼펙트론 대출 상담 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Application {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  memo?: string | null;
  createdAt: string;
}

export interface ApplicationInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  memo?: string;
}

export interface ApplicationListResponse {
  applications: Application[];
  total: number;
  page: number;
  limit: number;
}

export interface ApplicationStats {
  total: number;
  today: number;
  thisWeek: number;
  thisMonth: number;
}

export interface Settings {
  kakaoLink: string;
  telegramBotToken: string;
  telegramChatId: string;
  companyName: string;
  representative: string;
  businessNumber: string;
  address: string;
  phone: string;
}

export interface SettingsInput {
  kakaoLink?: string;
  telegramBotToken?: string;
  telegramChatId?: string;
  companyName?: string;
  representative?: string;
  businessNumber?: string;
  address?: string;
  phone?: string;
}

export interface AdminCredentials {
  password: string;
}

export interface AdminAuthResult {
  success: boolean;
}

export interface AdminSession {
  authenticated: boolean;
}

export type ListApplicationsParams = {
page?: number;
limit?: number;
};