# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 퍼펙트론 대출 상담 API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: applications
    description: Loan application operations
  - name: settings
    description: Admin settings operations
  - name: admin
    description: Admin authentication
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /applications:
    post:
      operationId: createApplication
      tags: [applications]
      summary: Submit loan consultation application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Invalid input
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all applications (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationListResponse"
        "401":
          description: Unauthorized

  /applications/stats:
    get:
      operationId: getApplicationStats
      tags: [applications]
      summary: Get application statistics (admin)
      responses:
        "200":
          description: Application stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationStats"
        "401":
          description: Unauthorized

  /applications/{id}:
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete an application (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get admin settings
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update admin settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401":
          description: Unauthorized

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResult"
        "401":
          description: Invalid credentials

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Unauthorized

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Application:
      type: object
      required: [id, name, phone, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        memo:
          type: ["string", "null"]
        createdAt:
          type: string

    ApplicationInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        memo:
          type: string

    ApplicationListResponse:
      type: object
      required: [applications, total, page, limit]
      properties:
        applications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    ApplicationStats:
      type: object
      required: [total, today, thisWeek, thisMonth]
      properties:
        total:
          type: integer
        today:
          type: integer
        thisWeek:
          type: integer
        thisMonth:
          type: integer

    Settings:
      type: object
      required: [kakaoLink, telegramBotToken, telegramChatId, companyName, representative, businessNumber, address, phone]
      properties:
        kakaoLink:
          type: string
        telegramBotToken:
          type: string
        telegramChatId:
          type: string
        companyName:
          type: string
        representative:
          type: string
        businessNumber:
          type: string
        address:
          type: string
        phone:
          type: string

    SettingsInput:
      type: object
      properties:
        kakaoLink:
          type: string
        telegramBotToken:
          type: string
        telegramChatId:
          type: string
        companyName:
          type: string
        representative:
          type: string
        businessNumber:
          type: string
        address:
          type: string
        phone:
          type: string

    AdminCredentials:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AdminAuthResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean