# lib/api-zod/src/generated/types/application.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 퍼펙트론 대출 상담 API
 * OpenAPI spec version: 0.1.0
 */

export interface Application {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  memo?: string | null;
  createdAt: string;
}