# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 퍼펙트론 대출 상담 API
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicationInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  memo?: string;
}