# lib/api-zod/src/generated/types/applicationListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 퍼펙트론 대출 상담 API
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from './application';

export interface ApplicationListResponse {
  applications: Application[];
  total: number;
  page: number;
  limit: number;
}