# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 퍼펙트론 대출 상담 API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  kakaoLink: string;
  telegramBotToken: string;
  telegramChatId: string;
  companyName: string;
  representative: string;
  businessNumber: string;
  address: string;
  phone: string;
}